latest push changes

+++ nips11/supertrain_models.m
@@ -1,4 +1,4 @@
-function supertrain_models(models,am)
+function supertrain_models(models)
 %Retrain ESVM models via VisRegression
 
 VOCinit;
@@ -15,7 +15,7 @@ myRandomize;
 rrr = randperm(length(models));
 
 final_directory = ...
-    sprintf('%s/%s-rereg/',...
+    sprintf('%s/%s-vregmine/',...
             VOCopts.localdir,...
             models{1}.models_name);
 
@@ -30,9 +30,10 @@ for indexi = 1:length(models)
   index = rrr(indexi);
 
   %index = 5;
-  
-  %index = 68;
-  
+  %index = 16;
+ 
+  %index = 68; 
+  %index = 38;
   m = models{index};
 
   filer = sprintf('%s/%s.%s.%d.mat',final_directory,...
@@ -41,9 +42,8 @@ for indexi = 1:length(models)
   if fileexists(filer) || (mymkdir_dist(filerlock) == 0)
     continue
   end
-  fprintf(1,'warn lock disabled\n');
+  %fprintf(1,'warn lock disabled\n');
   
-
   gridfiler = sprintf('%s/%s-%s.%05d.mat',...
                       BASEDIR,...
                       models{1}.cls,...
@@ -53,8 +53,7 @@ for indexi = 1:length(models)
   grid = load(gridfiler);
   cb = grid.coarse_boxes;
 
-  %   disp('hack load')
-  
+  %   disp('hack load')  
   if 0
     load mym.mat
   else
@@ -90,10 +89,6 @@ for indexi = 1:length(models)
     %save mym.mat m
   end
 
-
-
-
-
   %get initial hog
   %m.model.w = reshape(m.model.x(:,1) - mean(m.model.x(:,1)),...
   %                    m.model.hg_size);
@@ -104,77 +99,83 @@ for indexi = 1:length(models)
   %svids = cellfun2(@(x)x.model.svids,models);
   %svids = cat(2,svids{:});
   
-  NITER = 3;
+  NITER = 10;
   %xstart = m.model.x(:,1);
   
-  savem = m;
+  %savem = m;
   
+  keepers = nms_objid(m.model.svids,.2);
+  fprintf(1,'after nms keepers is %d\n',length(keepers));
+  m.model.svids = m.model.svids(keepers);
+  m.model.nsv = m.model.nsv(:,keepers);
   %print iteration 0
-  Isv1 = get_sv_stack(m,12);    
-  imwrite(Isv1,sprintf('%s/%s.%d_iter_I=%05d.png', ...
-                       final_directory,m.curid,...
-                       m.objectid,0),'png');
+  % Isv1 = get_sv_stack(m,12);    
+  % imwrite(Isv1,sprintf('%s/%s.%d_iter_I=%05d.png', ...
+  %                      final_directory,m.curid,...
+  %                      m.objectid,0),'png');
   
+  fgall = cat(1, get_pascal_bg('train',['-' m.cls]),...
+              get_pascal_bg('trainval',m.cls));
+   
+  %mining_queue = initialize_mining_queue(fg);
   for bbb = 1:NITER    
-
-    m.model.svids = savem.model.svids;
-    m.model.nsv = savem.model.nsv;
-    
+    %m.model.svids = savem.model.svids;
+    %m.model.nsv = savem.model.nsv;
     [aa,bb] = sort(m.model.w(:)'*m.model.nsv,'descend');
 
-    if 0 %bbb > 1
-         %START ENHANCE
-         
-         curids = cellfun(@(x)x.curid,m.model.svids(bb(1:10)));
-         clear fg;
-         for z = 1:length(curids)
-           fg{z} = sprintf(VOCopts.imgpath,sprintf('%06d',curids(z)));
-         end
-         
-         mining_queue = initialize_mining_queue(fg,1:length(fg));
-         
-         mining_params = get_default_mining_params;
-         mining_params.SAVE_SVS = 1;
-         mining_params.detection_threshold = -1.0;
-         mining_params.thresh = -1.0;
-         
-         mining_params.TOPK = 5;
-         mining_params.MAX_WINDOWS_BEFORE_SVM = 2000;
-         mining_params.MAX_IMAGES_BEFORE_SVM = 2000;
-         mining_params.NMS_MINES_OS = 1.0;
-         
-         curm = m;
-         
-         
-         %THIS HACK must be done on a copy (it is a hack to get more dets)
-         curm.model.b = curm.model.b - 0.5;
-         [hn, mining_queue, mining_stats] = ...
-             load_hn_fg({curm}, mining_queue, fg, mining_params);
-         
-         %%% END ENHANCE
-         
-         
-         
-         fprintf(1,'getting overlaps with gt\n');
-         tic
-         [Amaxos,Amaxind,Amaxclass] = ...
-             get_overlaps_with_gt(m,hn.objids{1});
-         
-         for q = 1:length(hn.objids{1})
-           hn.objids{1}{q}.maxos = Amaxos(q);
-           hn.objids{1}{q}.maxind = Amaxind(q);
-           hn.objids{1}{q}.maxclass = Amaxclass(q);
-         end
-         toc
-         
-         %Concatenate detections with model stuff
-         m.model.svids = cat(2,m.model.svids, hn.objids{1});
-         m.model.nsv   = cat(2,m.model.nsv, hn.xs{1});
-         
-         [a,b,c,d,indicator] = find_set_membership(m.model.svids,m.cls);
-         for i = 1:length(m.model.svids)
-           m.model.svids{i}.set = indicator(i);
-         end     
+    if 1 
+      %START ENHANCE      
+      curids = cellfun(@(x)x.curid,m.model.svids(bb(1:50)));
+      clear fg
+      fg = cell(length(curids),1);
+      for z = 1:length(curids)
+        fg{z} = sprintf(VOCopts.imgpath,sprintf('%06d',curids(z)));
+      end
+      rrr2 = randperm(length(fgall));
+      fg = cat(1,fg,fgall(rrr2(1:50)));
+      fg = unique(fg);
+      mining_queue = initialize_mining_queue(unique(fg));
+      %mining_queue = mining_queue(1:10);
+      
+      mining_params = get_default_mining_params;
+      mining_params.SAVE_SVS = 1;
+      mining_params.detection_threshold = -1.0;
+      mining_params.thresh = -1.0;
+      
+      mining_params.TOPK = 5;
+      mining_params.MAX_WINDOWS_BEFORE_SVM = 200;
+      mining_params.MAX_IMAGES_BEFORE_SVM = 100;
+      mining_params.NMS_MINES_OS = 1.0;
+      
+      curm = m;
+      
+      %THIS HACK must be done on a copy (it is a hack to get more dets)
+      curm.model.b = curm.model.b - 0.1;
+      [hn, mining_queue, mining_stats] = ...
+          load_hn_fg({curm}, mining_queue, fg, mining_params);
+      
+      %%% END ENHANCE
+        
+      fprintf(1,'getting overlaps with gt\n');
+      tic
+      [Amaxos,Amaxind,Amaxclass] = ...
+          get_overlaps_with_gt(m,hn.objids{1});
+      
+      for q = 1:length(hn.objids{1})
+        hn.objids{1}{q}.maxos = Amaxos(q);
+        hn.objids{1}{q}.maxind = Amaxind(q);
+        hn.objids{1}{q}.maxclass = Amaxclass(q);
+      end
+      toc
+      
+      %Concatenate detections with model stuff
+      m.model.svids = cat(2,m.model.svids, hn.objids{1});
+      m.model.nsv   = cat(2,m.model.nsv, hn.xs{1});
+      
+      [a,b,c,d,indicator] = find_set_membership(m.model.svids,m.cls);
+      for i = 1:length(m.model.svids)
+        m.model.svids{i}.set = indicator(i);
+      end     
     end
     
     [aa,bb] = sort(m.model.w(:)'*m.model.nsv,'descend');
@@ -187,19 +188,20 @@ for indexi = 1:length(models)
     m.model.svids = m.model.svids(keepers);
     m.model.nsv = m.model.nsv(:,keepers);
     m = cap_to_K_dets(m,2000);
-    %m.model.xstart = xstart;
     
+    
+    mold = m;
     m = do_rank(m,mining_params);
     
-    Isv1 = get_sv_stack(m,12);
-    imwrite(Isv1,sprintf('%s/%s.%d_iter_I=%05d.png', ...
-                         final_directory,m.curid,...
-                         m.objectid,bbb),'png');
+    % mold.model.svids = m.model.svids;
+    % mold.model.nsv = m.model.nsv;
+    
+    % Isv1 = get_sv_stack(mold,12);
+    % imwrite(Isv1,sprintf('%s/%s.%d_iter_I=%05d-oldw.png', ...
+    %                      final_directory,m.curid,...
+    %                      m.objectid,bbb),'png');
+    
     figure(1)
-    clf
-    imagesc(Isv1)
-    drawnow
-  
     show_cool_os(m)
     drawnow
     
@@ -210,11 +212,12 @@ for indexi = 1:length(models)
       print(gcf,sprintf('%s/%s.%d_iter=%05d.png', ...
                         final_directory,m.curid,...
                         m.objectid,bbb),'-dpng'); 
+      
+      Isv1 = get_sv_stack(m,12);
+      imwrite(Isv1,sprintf('%s/%s.%d_iter_I=%05d.png', ...
+                           final_directory,m.curid,...
+                           m.objectid,bbb),'png');
     end
-
-         
-    
-    
            
     if bbb == NITER
       break;
@@ -265,7 +268,9 @@ for indexi = 1:length(models)
   
   %m.models_name = [m.models_name '-rereg'];
   save(filer,'m');
-  rmdir(filerlock);
+  if exists(filerlock,'dir')
+    rmdir(filerlock);
+  end
 end
